+++ src/OpenTelemetry.Instrumentation.StackExchangeRedis/StackExchangeRedisConnectionInstrumentation.cs
@@ -31,9 +31,9 @@ internal sealed class StackExchangeRedisConnectionInstrumentation : IDisposable
 {
     internal const string RedisDatabaseIndexKeyName = "db.redis.database_index";
     internal const string RedisFlagsKeyName = "db.redis.flags";
-    internal static readonly string ActivitySourceName = typeof(StackExchangeRedisConnectionInstrumentation).Assembly.GetName().Name;
+    internal static readonly string ActivitySourceName = typeof(StackExchangeRedisConnectionInstrumentation).Assembly.GetName().Name!;
     internal static readonly string ActivityName = ActivitySourceName + ".Execute";
-    internal static readonly Version Version = typeof(StackExchangeRedisConnectionInstrumentation).Assembly.GetName().Version;
+    internal static readonly Version Version = typeof(StackExchangeRedisConnectionInstrumentation).Assembly.GetName().Version!;
     internal static readonly ActivitySource ActivitySource = new(ActivitySourceName, Version.ToString());
     internal static readonly IEnumerable<KeyValuePair<string, object?>> CreationTags = new[]
     {
@@ -137,7 +137,7 @@ internal sealed class StackExchangeRedisConnectionInstrumentation : IDisposable
         }
     }
 
-    private void DrainEntries(object state)
+    private void DrainEntries(object? state)
     {
         while (true)
         {